+++ EquipmentAndQuickSlots/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Release 2.1.12
 * Updated for Valheim 0.217.27
 * Updated Auga and CLLC API's
+* If Auga is loaded, set the default position of Quick Slot Bar accordingly
 
 <details>
 <summary><b>Changelog History</b> (<i>click to expand</i>)</summary>